[refactor] tmap 패키지 global -> direction으로 이동
