+++ src/main/java/com/wayble/server/ServerApplication.java
@@ -1,6 +1,6 @@
 package com.wayble.server;
 
-import com.wayble.server.common.client.tmap.TMapProperties;
+import com.wayble.server.direction.external.tmap.TMapProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.elasticsearch.ReactiveElasticsearchRepositoriesAutoConfiguration;
