+++ src/main/java/com/wayble/server/direction/external/tmap/TMapProperties.java
@@ -1,4 +1,4 @@
-package com.wayble.server.common.client.tmap;
+package com.wayble.server.direction.external.tmap;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
